+++ xdef/src/main/java/org/xdef/component/XCEnumeration.java
@@ -1,6 +1,8 @@
 package org.xdef.component;
 
-public interface XCEnumeration {
+import java.io.Serializable;
+
+public interface XCEnumeration extends Serializable {
 
 	/** Get object associated with this item of enumeration.
 	 * @return object associated with this item of enumeration.