+++ xdef/src/test/java/test/xdef/TestXComponents_Y04.java
@@ -1,6 +1,8 @@
 package test.xdef;
 
-public class TestXComponents_Y04 {
+import java.io.Serializable;
+
+public class TestXComponents_Y04 implements Serializable {
 	private String _One;
 	public String getJedna() {return _One;}
 	public void setJedna(String One) {_One = One.equals("1") ? "One" : One;}