+++ xdef/src/test/java/test/xdef/TestXComponents_Y06Container.java
@@ -1,7 +1,9 @@
 package test.xdef;
 
+import java.io.Serializable;
+
 public class TestXComponents_Y06Container <T extends TestXComponents_Y06Domain>
-implements TestXComponents_Y06DomainContainer<T> {
+implements TestXComponents_Y06DomainContainer<T>, Serializable {
 	T domain;
 
 	@Override